user added

+++ src/main/java/ru/anikanov/tm/command/project/ProjectDeleteCommand.java
@@ -2,9 +2,6 @@ package ru.anikanov.tm.command.project;
 
 import ru.anikanov.tm.Bootstrap;
 import ru.anikanov.tm.command.AbstractCommand;
-import ru.anikanov.tm.service.ProjectService;
-import ru.anikanov.tm.service.TaskService;
-
 import java.text.ParseException;
 
 public class ProjectDeleteCommand extends AbstractCommand {
@@ -22,7 +19,9 @@ public class ProjectDeleteCommand extends AbstractCommand {
         return "command to delete project";
     }
     @Override
-    public void execute(String name) throws ParseException {
+    public void execute() throws ParseException {
+        if (!isSecure()) return;
+        String name = scanner.next();
         bootstrap.projectService.remove(name);
     }
 }
